Configuracao inicial do xinit e OpenBox

+++ install.sh
@@ -11,20 +11,27 @@ sudo reflector --verbose --country Brazil --score 100 --fastest 50 --sort rate -
 sudo pacman -Syyu base-devel git --needed --noconfirm
 
 # instalar o yay
-git clone https://aur.archlinux.org/yay.git
-cd yay
-makepkg -si --noconfirm
-cd ..
-rm -rf yay
+if pacman -Qs yay > /dev/null ; then
+   echo "yay já instalado"
+else
+   git clone https://aur.archlinux.org/yay.git
+   cd yay
+   makepkg -si --noconfirm
+   cd ..
+   rm -rf yay
+fi
 
 # instalar pacotes
 sudo pacman -S vim net-tools zip unzip tar unrar curl wget cronie xorg xorg-xinit openbox thunar xterm slim tint2 networkmanager network-manager-applet ttf-dejavu ttf-liberation --needed --noconfirm
 
 # configurar o xinit
-# cp config/_xinitrc ~/.xinitrc
+cp files/home/user/_xinitrc ~/.xinitrc
+chmod +x ~/.xinitrc
 
 # configurar o openbox
 mkdir -p ~/.config/openbox
+cp files/home/user/_config/openbox/* ~/.config/openbox/
 
-# habilitar do slim
+# habilitar o slim
 sudo systemctl enable slim
+
